+++ config.py
@@ -3,6 +3,9 @@ INPUT_GFF3 = "./test_data/gff3_folder"
 INPUT_ORF = "./test_data/orf_folder"
 OUTPUT_FOLDER = "./lumbridge_output"
 
+# features to extract from gff3
+GFF3_FEATURES = ["miRNA", "rRNA", "gene"]
+
 # Homer2
 HOMER2_BIN_PATH = "/usr/src/app/bin"
 HOMER2_OUTPUT_FOLDER = "./lumbridge_output/homer2"