+++ lumbridge/__init__.py
@@ -91,38 +91,30 @@ def run(test_config: str | None = None) -> None:
     )
 
     log.info("------ Extract gff3 positive elements  -------")
-    gff3_folder_pos_strand: str = (
-        f"{parsed_config['OUTPUT_FOLDER']}/gff3_positive_strand"
-    )
+    gff3_folder_pos_strand: str = f"{parsed_config['OUTPUT_FOLDER']}/gff3_positive_strand"
     extract_positive_elements_gff3(parsed_config["INPUT_GFF3"], gff3_folder_pos_strand)
 
     log.info("------ Create file with ORF only on forward strand  -------")
-    orf_output_folder: str = (
-        f"{parsed_config['OUTPUT_FOLDER']}/orf_folder_positive_strand"
-    )
+    orf_output_folder: str = f"{parsed_config['OUTPUT_FOLDER']}/orf_folder_positive_strand"
     create_output_orf_forward_strand(parsed_config["INPUT_ORF"], orf_output_folder)
 
     log.info("------ Find ORf overlaps in GFF3 file  -------")
     orf_in_gff3_folder: str = f"{parsed_config['OUTPUT_FOLDER']}/orf_in_gff3"
-    find_gff3_and_orf_intervals(
-        orf_output_folder, gff3_folder_pos_strand, orf_in_gff3_folder
-    )
+    find_gff3_and_orf_intervals(orf_output_folder, gff3_folder_pos_strand, orf_in_gff3_folder)
 
     log.info("------ Find Polyadenylation sequences in Fasta file  -------")
-    find_poly_adi_sequences(
-        parsed_config["INPUT_FASTA"], parsed_config["OUTPUT_FOLDER"]
-    )
+    find_poly_adi_sequences(parsed_config["INPUT_FASTA"], parsed_config["OUTPUT_FOLDER"])
 
     log.info("------ Make homer2 output -------")
-    make_homer2_output(
-        parsed_config["HOMER2_BIN_PATH"],
-        parsed_config["INPUT_FASTA"],
-        gff3_folder_pos_strand,
-        parsed_config["HOMER2_OUTPUT_FOLDER"],
-        parsed_config["HOMER2_UPSTREAM_GEN_SEQ_LENGTH"],
-        parsed_config["HOMER2_DOWNSTREAM_GEN_SEQ_LENGTH"],
-        cpu_cores=parsed_config["HOMER2_CPU_CORES"],
-    )
+    # make_homer2_output(
+    #     parsed_config["HOMER2_BIN_PATH"],
+    #     parsed_config["INPUT_FASTA"],
+    #     gff3_folder_pos_strand,
+    #     parsed_config["HOMER2_OUTPUT_FOLDER"],
+    #     parsed_config["HOMER2_UPSTREAM_GEN_SEQ_LENGTH"],
+    #     parsed_config["HOMER2_DOWNSTREAM_GEN_SEQ_LENGTH"],
+    #     cpu_cores=parsed_config["HOMER2_CPU_CORES"],
+    # )
 
     log.info("------ Annotate homer2 output to fasta -------")
     annotate_homer2_motifs(
@@ -144,6 +136,7 @@ def run(test_config: str | None = None) -> None:
         poly_adenyl_folder=f"{parsed_config['OUTPUT_FOLDER']}/poly_adenylation",
         output_folder=f"{parsed_config['OUTPUT_FOLDER']}/model_data",
         max_feature_overlap=parsed_config["MAX_FEATURE_OVERLAP"],
+        gff3_features=parsed_config["GFF3_FEATURES"],
         version=__version__,
     )